app/requirements.txt
fastapi
python-dotenv
apscheduler
uvicorn